[front] - refacto(members): workspace members (#7334)
* [front/components/workspace] - feature: implement ChangeMemberModal functionality

 - Add modal to handle workspace member role updates and access revocation
 - Utilize SWR for data mutation after member role changes or access revocation

* [front/lib/api] - feature: extend searchMembers API to include workspace roles

 - Modified searchMembers function to return UserTypeWithWorkspaces, enriching user objects with their workspace roles
 - Added logic to fetch memberships by user IDs and map them to corresponding users adding their roles to the response

[front/lib/resources] - feature: implement fetchByUserIds for membership retrieval

 - Introduced a new method fetchByUserIds in MembershipResource to retrieve memberships based on an array of user IDs

[front/pages/api] - refactor: update API response type to include workspaces in member search

 - Modified the response type of SearchMembers API to use UserTypeWithWorkspaces, reflecting changes in the searchMembers function

* [front/components/members] - feature: revamp MembersList with search and pagination

- Implement a new DataTable component to handle member listing with built-in search and pagination support
- Replace static members rendering with dynamic rows generation based on new data fetching logic
- Add functionality to invite members to the workspace and handle member role changes with a modal
- Remove explicit sorting and filtering, delegating these responsibilities to the DataTable component

* [members] - refactor: streamline member invitation process

 - Remove direct members list dependency in InviteEmailModal and fetch member information asynchronously when required
 - Adjust logic to work with fetched member data and improve error handling for member information requests
 - Implement a cleaner separation of concerns by removing the members prop and handling member existence checks internally

[MembersList] - refactor: clean up and standardize formatting of MembersList code

 - Standardize formatting for improved code consistency and readability
 - Remove unused imports and code related to the member role change functionality, which has been refactored elsewhere

[CreateOrEditVaultModal] - devops: add console logging for debugging member data

 - Temporary log statement added to help in debugging during development phase

[workspace] - refactor: simplify role assignment logic in workspace API

 - Refactor role determination code for cleaner and more succinct expression
 - Use type imports where necessary to streamline the codebase

[membership_resource] - refactor: remove unused import from membership resource file

 - Clean up unused User model import to keep the codebase lean

[workspace/members] - refactor: simplify WorkspaceAdmin component

 - Remove nested MemberList component and integrate member list logic directly into WorkspaceAdmin component
 - Simplify popup logic associated with inviting members by utilizing ReachLimitPopup component to centralize the display logic
 - Clean up and remove commented code and redundant imports resulting in a more concise and readable component structure

* [front] - refactor: remove console.log from CreateOrEditVaultModal

 - Cleaned up console logging to comply with best practices for production code

* [front/components/members] - refactor: streamline MembersList and search functionality

 - Remove searchbar and invite button from MembersList component
 - Shift search and invitation UI to parent component for better separation of concerns

[front/pages/w/[wId]/members] - feature: integrate search functionality with MembersList and InvitationsList

 - Add a search bar to filter both members and invitations by email
 - Include an 'Invite members' button with functionality to open invitation modal
 - Pass down search term to MembersList and InvitationsList components to filter displayed items

* [front/components/members] - refactor: extract EditInvitationModal to its own component

 - Moved EditInvitationModal from InvitationModal.tsx to EditInvitationModal.tsx to separate concerns
 - Updated import paths in InvitationsList.tsx to reflect the relocation of EditInvitationModal

[front/lib] - feature: implement invitation update and resend logic

 - Added updateInvitation function to handle invitation role updates and revocation
 - Created sendInvitations function to manage sending new invitations and resending existing ones with proper error handling

[front/components/members] - refactor: clean up InvitationModal.tsx

 - Removed EditInvitationModal related code due to its extraction into a separate file
 - Cleaned up unused import statements and functions, simplifying the InvitationModal component

[front/lib] - refactor: move MAX_UNCONSUMED_INVITATIONS_PER_WORKSPACE_PER_DAY constant to invitations.ts

 - Centralized the invitation limit constant for better maintainability

* [front/components/members] - feature: set explicit column widths in MembersList component

 - Define specific width for 'name' and 'role' columns to ensure consistent presentation

[front] - fix: update @dust-tt/sparkle package to version 0.2.238

 - Upgrade @dust-tt/sparkle to incorporate the latest fixes and enhancements

* [front/components/members] - refactor: overhaul InvitationsList with DataTable component

 - Replace manual rendering of invitations list with DataTable component for better performance and maintainability
 - Use useMemo for filtering and sorting invitations to optimize re-rendering
 - Add onClick logic to row data for handling invitation selection
 - Adjust column width in MembersList for consistency with new DataTable layout

* [front] - refactor: update type of CellContext to include string data

 - Changed the second generic parameter of CellContext from unknown to string to explicitly type the data associated with table cells

* [front/components/members] - refactor: update role handling in MembersList component

 - Replace ActiveRoleType with more generic RoleType in type definitions
 - Introduce displayRole utility for consistent role display logic
 - Adjust the role label and color display to handle a new "none" role condition
 - Import and utilize the displayRole function from Roles component

* [front/components/members] - refactor: update MembersList to filter out users with 'none' role

 - Apply filtering logic to exclude members with 'none' roles from the members list
 - Enhance efficiency by recalculating the filtered list only when the list of members changes

* add debug logging

* add more debug logging

* [front] - fix: ensure member search is within the correct workspace

 - Filter out memberships that do not belong to the workspace owned by the requesting user
 - Previously, membership search could return results from outside the intended workspace, which has now been corrected

* [front/components/members] - refactor: clean up console logging in MembersList component

 - Remove unnecessary console logs to declutter console output

[front/lib/api] - refactor: switch to `getActiveMemberships` for member fetching

 - Use new `getActiveMemberships` method to directly obtain active memberships for better clarity and potentially improved performance

* [front/members] - refactor: update member invitation and listing logic to use workspace roles

 - Modify the invitation modal to check user roles within workspaces instead of directly on the user object
 - Adjust the members list filter to only show members with roles other than "none" in the workspace context
 - Update the check for not in workspace members by using workspace-specific role information in the InvitationModal

* [front] - fix: correct member invitation check conditions

 - Ensure non-existing workspace members are filtered properly
 - Handle cases where existing members may not have workspace roles

---------
