+++ front/components/members/InvitationsList.tsx
@@ -1,10 +1,22 @@
-import { Avatar, ChevronRightIcon, Chip, Icon, Page } from "@dust-tt/sparkle";
+import {
+  Avatar,
+  ChevronRightIcon,
+  Chip,
+  DataTable,
+  Icon,
+  Page,
+} from "@dust-tt/sparkle";
 import type { MembershipInvitationType, WorkspaceType } from "@dust-tt/types";
-import { useState } from "react";
+import type { CellContext } from "@tanstack/react-table";
+import React, { useMemo, useState } from "react";
 
-import { EditInvitationModal } from "@app/components/members/InvitationModal";
+import { EditInvitationModal } from "@app/components/members/EditInvitationModal";
 import { displayRole, ROLES_DATA } from "@app/components/members/Roles";
 import { useWorkspaceInvitations } from "@app/lib/swr/memberships";
+
+type RowData = MembershipInvitationType & {
+  onClick: () => void;
+};
 
 export function InvitationsList({
   owner,
@@ -16,14 +28,68 @@ export function InvitationsList({
   const { invitations, isInvitationsLoading } = useWorkspaceInvitations(owner);
   const [selectedInvite, setSelectedInvite] =
     useState<MembershipInvitationType | null>(null);
-  const filteredInvitations = invitations
-    .sort((a, b) => a.inviteEmail.localeCompare(b.inviteEmail))
-    .filter((i) => i.status === "pending")
-    .filter(
-      (i) =>
-        !searchText ||
-        i.inviteEmail.toLowerCase().includes(searchText.toLowerCase())
-    );
+
+  const filteredInvitations = useMemo(
+    () =>
+      invitations
+        .sort((a, b) => a.inviteEmail.localeCompare(b.inviteEmail))
+        .filter((i) => i.status === "pending")
+        .filter(
+          (i) =>
+            !searchText ||
+            i.inviteEmail.toLowerCase().includes(searchText.toLowerCase())
+        ),
+    [invitations, searchText]
+  );
+
+  const rows = useMemo(
+    () =>
+      filteredInvitations.map((invitation) => ({
+        ...invitation,
+        onClick: () => setSelectedInvite(invitation),
+      })),
+    [filteredInvitations]
+  );
+
+  const columns = [
+    {
+      id: "inviteEmail",
+      cell: (info: CellContext<RowData, string>) => (
+        <DataTable.CellContent>
+          <span>{info.row.original.inviteEmail}</span>
+        </DataTable.CellContent>
+      ),
+      meta: {
+        width: "46rem",
+      },
+    },
+    {
+      id: "initialRole",
+      cell: (info: CellContext<RowData, string>) => (
+        <DataTable.CellContent>
+          <Chip
+            size="xs"
+            color={ROLES_DATA[info.row.original.initialRole]["color"]}
+            className="capitalize"
+          >
+            {displayRole(info.row.original.initialRole)}
+          </Chip>
+        </DataTable.CellContent>
+      ),
+      meta: {
+        width: "6rem",
+      },
+    },
+    {
+      id: "action",
+      cell: () => (
+        <DataTable.CellContent>
+          <Icon visual={ChevronRightIcon} className="text-element-600" />
+        </DataTable.CellContent>
+      ),
+    },
+  ];
+
   return (
     <>
       {selectedInvite && (
@@ -36,41 +102,9 @@ export function InvitationsList({
       {filteredInvitations.length > 0 && (
         <Page.H variant="h5">Invitations</Page.H>
       )}
-      <div className="s-w-full">
-        {!isInvitationsLoading &&
-          filteredInvitations.length > 0 &&
-          filteredInvitations.map((invitation: MembershipInvitationType) => (
-            <div
-              key={`invitation-${invitation.id}`}
-              className="transition-color flex cursor-pointer items-center justify-center gap-3 border-t border-structure-200 p-2 text-xs duration-200 hover:bg-action-50 sm:text-sm"
-              onClick={() => setSelectedInvite(invitation)}
-            >
-              <div className="hidden sm:block">
-                <Avatar size="sm" className={invitation.sId} />
-              </div>
-              <div className="flex grow flex-col gap-1 sm:flex-row sm:gap-3">
-                <div className="grow font-normal text-element-700">
-                  {invitation.inviteEmail}
-                </div>
-                <div>
-                  <Chip
-                    size="xs"
-                    color={ROLES_DATA[invitation.initialRole]["color"]}
-                    className="capitalize"
-                  >
-                    {displayRole(invitation.initialRole)}
-                  </Chip>
-                </div>
-                <div className="hidden sm:block">
-                  <Icon
-                    visual={ChevronRightIcon}
-                    className="text-element-600"
-                  />
-                </div>
-              </div>
-            </div>
-          ))}
-        {isInvitationsLoading && (
+
+      {isInvitationsLoading ? (
+        <div className="flex flex-col gap-2">
           <div className="flex animate-pulse cursor-pointer items-center justify-center gap-3 border-t border-structure-200 bg-structure-50 py-2 text-xs sm:text-sm">
             <div className="hidden sm:block">
               <Avatar size="xs" />
@@ -88,8 +122,12 @@ export function InvitationsList({
               <ChevronRightIcon />
             </div>
           </div>
-        )}
-      </div>
+        </div>
+      ) : (
+        filteredInvitations.length > 0 && (
+          <DataTable data={rows} columns={columns} />
+        )
+      )}
     </>
   );
 }
