+++ front/lib/api/workspace.ts
@@ -3,12 +3,12 @@ import type {
   MembershipRoleType,
   RoleType,
   SubscriptionType,
-  UserType,
   UserTypeWithWorkspaces,
   WorkspaceDomain,
   WorkspaceSegmentationType,
   WorkspaceType,
 } from "@dust-tt/types";
+import { ACTIVE_ROLES } from "@dust-tt/types";
 
 import type { PaginationParams } from "@app/lib/api/pagination";
 import type { Authenticator } from "@app/lib/auth";
@@ -184,7 +184,7 @@ export async function searchMembers(
     email?: string;
   },
   paginationParams: PaginationParams
-): Promise<{ members: UserType[]; total: number }> {
+): Promise<{ members: UserTypeWithWorkspaces[]; total: number }> {
   const owner = auth.workspace();
   if (!owner) {
     return { members: [], total: 0 };
@@ -198,7 +198,29 @@ export async function searchMembers(
     paginationParams
   );
 
-  return { members: users.map((u) => u.toJSON()), total };
+  const { memberships } = await MembershipResource.getActiveMemberships({
+    users,
+    workspace: owner,
+  });
+
+  const usersWithWorkspaces = users.map((u) => {
+    const membership = memberships.find(
+      (m) => m.userId === u.id && m.workspaceId === owner.id
+    );
+    const role =
+      membership && !membership.isRevoked()
+        ? ACTIVE_ROLES.includes(membership.role)
+          ? membership.role
+          : ("none" as RoleType)
+        : ("none" as RoleType);
+
+    return {
+      ...u.toJSON(),
+      workspaces: [{ ...owner, role, flags: null }],
+    };
+  });
+
+  return { members: usersWithWorkspaces, total };
 }
 
 export async function getMembersCount(
