+++ front/pages/w/[wId]/members/index.tsx
@@ -1,39 +1,27 @@
 import {
-  Avatar,
   Button,
   Dialog,
-  ElementModal,
   Page,
   PlusIcon,
   Popup,
   Searchbar,
 } from "@dust-tt/sparkle";
-import type {
-  ActiveRoleType,
-  UserType,
-  UserTypeWithWorkspaces,
-  WorkspaceDomain,
-  WorkspaceType,
-} from "@dust-tt/types";
+import type { UserType, WorkspaceDomain, WorkspaceType } from "@dust-tt/types";
 import type {
   PlanType,
   SubscriptionPerSeatPricing,
   SubscriptionType,
 } from "@dust-tt/types";
-import { isActiveRoleType } from "@dust-tt/types";
 import { UsersIcon } from "@heroicons/react/20/solid";
 import type { InferGetServerSidePropsType } from "next";
 import { useRouter } from "next/router";
-import { useContext, useMemo, useState } from "react";
-import { useSWRConfig } from "swr";
+import React, { useContext, useMemo, useState } from "react";
 
 import type { WorkspaceLimit } from "@app/components/app/ReachedLimitPopup";
 import { ReachedLimitPopup } from "@app/components/app/ReachedLimitPopup";
 import { InviteEmailModal } from "@app/components/members/InvitationModal";
 import { InvitationsList } from "@app/components/members/InvitationsList";
 import { MembersList } from "@app/components/members/MembersList";
-import { ROLES_DATA } from "@app/components/members/Roles";
-import { RoleDropDown } from "@app/components/members/RolesDropDown";
 import { subNavigationAdmin } from "@app/components/navigation/config";
 import AppLayout from "@app/components/sparkle/AppLayout";
 import { SendNotificationsContext } from "@app/components/sparkle/Notification";
@@ -45,11 +33,9 @@ import {
   checkWorkspaceSeatAvailabilityUsingAuth,
   getWorkspaceVerifiedDomain,
 } from "@app/lib/api/workspace";
-import { handleMembersRoleChange } from "@app/lib/client/members";
 import { withDefaultUserAuthRequirements } from "@app/lib/iam/session";
 import { isUpgraded } from "@app/lib/plans/plan_codes";
 import { getPerSeatSubscriptionPricing } from "@app/lib/plans/subscription";
-import { useMembers } from "@app/lib/swr/memberships";
 
 export const getServerSideProps = withDefaultUserAuthRequirements<{
   user: UserType;
@@ -109,12 +95,44 @@ export default function WorkspaceAdmin({
   workspaceHasAvailableSeats,
 }: InferGetServerSidePropsType<typeof getServerSideProps>) {
   const router = useRouter();
+  const [searchTerm, setSearchTerm] = useState("");
   const [showNoInviteLinkPopup, setShowNoInviteLinkPopup] = useState(false);
   const [isActivateAutoJoinOpened, setIsActivateAutoJoinOpened] =
     useState(false);
+  const [inviteEmailModalOpen, setInviteEmailModalOpen] = useState(false);
+  const [inviteBlockedPopupReason, setInviteBlockedPopupReason] =
+    useState<WorkspaceLimit | null>(null);
 
   const { domain = "", domainAutoJoinEnabled = false } =
     workspaceVerifiedDomain ?? {};
+
+  const onInviteClick = () => {
+    if (!isUpgraded(plan)) {
+      setInviteBlockedPopupReason("cant_invite_free_plan");
+    } else if (subscription.paymentFailingSince) {
+      setInviteBlockedPopupReason("cant_invite_payment_failure");
+    } else if (!workspaceHasAvailableSeats) {
+      setInviteBlockedPopupReason("cant_invite_no_seats_available");
+    } else {
+      setInviteEmailModalOpen(true);
+    }
+  };
+
+  const popup = useMemo(() => {
+    if (!inviteBlockedPopupReason) {
+      return <></>;
+    }
+
+    return (
+      <ReachedLimitPopup
+        isOpened={!!inviteBlockedPopupReason}
+        onClose={() => setInviteBlockedPopupReason(null)}
+        subscription={subscription}
+        owner={owner}
+        code={inviteBlockedPopupReason}
+      />
+    );
+  }, [inviteBlockedPopupReason, owner, subscription]);
 
   return (
     <AppLayout
@@ -194,112 +212,40 @@ export default function WorkspaceAdmin({
           plan={plan}
           strategyDetails={enterpriseConnectionStrategyDetails}
         />
-        <MemberList perSeatPricing={perSeatPricing} />
-      </Page.Vertical>
-    </AppLayout>
-  );
-
-  function MemberList({
-    perSeatPricing,
-  }: {
-    perSeatPricing: SubscriptionPerSeatPricing | null;
-  }) {
-    const [inviteBlockedPopupReason, setInviteBlockedPopupReason] =
-      useState<WorkspaceLimit | null>(null);
-
-    const [searchText, setSearchText] = useState("");
-    const { members, isMembersLoading } = useMembers(owner);
-
-    const [inviteEmailModalOpen, setInviteEmailModalOpen] = useState(false);
-
-    const [changeRoleMember, setChangeRoleMember] =
-      useState<UserTypeWithWorkspaces | null>(null);
-
-    const popup = useMemo(() => {
-      if (!inviteBlockedPopupReason) {
-        return <></>;
-      }
-
-      return (
-        <ReachedLimitPopup
-          isOpened={!!inviteBlockedPopupReason}
-          onClose={() => setInviteBlockedPopupReason(null)}
-          subscription={subscription}
+        <div className="flex flex-row gap-2">
+          <Searchbar
+            placeholder="Search members (email)"
+            value={searchTerm}
+            name="search"
+            onChange={(s) => {
+              setSearchTerm(s);
+            }}
+          />
+          <Button
+            label="Invite members"
+            icon={PlusIcon}
+            onClick={onInviteClick}
+          />
+        </div>
+        <InvitationsList owner={owner} searchText={searchTerm} />
+        <MembersList
+          currentUserId={user.sId}
           owner={owner}
-          code={inviteBlockedPopupReason}
+          searchText={searchTerm}
         />
-      );
-    }, [inviteBlockedPopupReason, setInviteBlockedPopupReason]);
-    return (
-      <>
         <InviteEmailModal
           showModal={inviteEmailModalOpen}
           onClose={() => {
             setInviteEmailModalOpen(false);
           }}
           owner={owner}
-          prefillText={searchText}
-          members={members}
+          prefillText={""}
           perSeatPricing={perSeatPricing}
         />
-        <ChangeMemberModal
-          owner={owner}
-          member={changeRoleMember}
-          onClose={() => setChangeRoleMember(null)}
-        />
-        <Page.Vertical gap="sm" align="stretch">
-          <Page.H variant="h5">Member list</Page.H>
-          <div className="flex flex-col items-stretch gap-2 sm:flex-row">
-            <div className="flex-grow">
-              <Searchbar
-                placeholder="Search members"
-                onChange={setSearchText}
-                value={searchText}
-                name={""}
-              />
-            </div>
-            <div className="relative flex-none">
-              <Button
-                variant="primary"
-                label="Invite members"
-                size="sm"
-                icon={PlusIcon}
-                onClick={() => {
-                  if (!isUpgraded(plan)) {
-                    setInviteBlockedPopupReason("cant_invite_free_plan");
-                  } else if (subscription.paymentFailingSince) {
-                    setInviteBlockedPopupReason("cant_invite_payment_failure");
-                  } else if (!workspaceHasAvailableSeats) {
-                    setInviteBlockedPopupReason(
-                      "cant_invite_no_seats_available"
-                    );
-                  } else {
-                    setInviteEmailModalOpen(true);
-                  }
-                }}
-              />
-              {popup}
-            </div>
-          </div>
-          <div className="s-w-full">
-            <div className="space-y-2 pt-4">
-              <InvitationsList owner={owner} searchText={searchText} />
-            </div>
-            <div className="space-y-2 pb-3 pt-4">
-              <Page.H variant="h5">Members</Page.H>
-              <MembersList
-                users={members}
-                currentUserId={user.id}
-                isMembersLoading={isMembersLoading}
-                onClickEvent={(role) => setChangeRoleMember(role)}
-                searchText={searchText}
-              />
-            </div>
-          </div>
-        </Page.Vertical>
-      </>
-    );
-  }
+        {popup}
+      </Page.Vertical>
+    </AppLayout>
+  );
 }
 
 function DomainAutoJoinModal({
@@ -370,124 +316,5 @@ function DomainAutoJoinModal({
     >
       <div>{description}</div>
     </Dialog>
-  );
-}
-
-function ChangeMemberModal({
-  onClose,
-  member,
-  owner,
-}: {
-  onClose: () => void;
-  member: UserTypeWithWorkspaces | null;
-  owner: WorkspaceType;
-}) {
-  const { role = null } = member?.workspaces[0] ?? {};
-
-  const { mutate } = useSWRConfig();
-  const sendNotification = useContext(SendNotificationsContext);
-  const [revokeMemberModalOpen, setRevokeMemberModalOpen] = useState(false);
-  const [selectedRole, setSelectedRole] = useState<ActiveRoleType | null>(
-    role !== "none" ? role : null
-  );
-  const [isSaving, setIsSaving] = useState(false);
-
-  if (!member || !role || !isActiveRoleType(role)) {
-    return null;
-  }
-
-  return (
-    <ElementModal
-      openOnElement={member}
-      onClose={() => {
-        onClose();
-        setSelectedRole(null);
-        setIsSaving(false);
-      }}
-      isSaving={isSaving}
-      hasChanged={selectedRole !== member.workspaces[0].role}
-      title={member.fullName || "Unreachable"}
-      variant="side-sm"
-      onSave={async (closeModalFn: () => void) => {
-        if (!selectedRole) {
-          return;
-        }
-        setIsSaving(true);
-        await handleMembersRoleChange({
-          members: [member],
-          role: selectedRole,
-          sendNotification,
-        });
-        await mutate(`/api/w/${owner.sId}/members`);
-        closeModalFn();
-      }}
-      saveLabel="Update role"
-    >
-      <Page variant="modal">
-        <div className="mt-6 flex flex-col gap-9 text-sm text-element-700">
-          <div className="flex items-center gap-4">
-            <Avatar size="lg" visual={member.image} name={member.fullName} />
-            <div className="flex grow flex-col">
-              <div className="font-semibold text-element-900">
-                {member.fullName}
-              </div>
-              <div className="font-normal">{member.email}</div>
-            </div>
-          </div>
-          <div className="flex flex-col gap-2">
-            <div className="flex items-center gap-2">
-              <div className="font-bold text-element-900">Role:</div>
-              <RoleDropDown
-                selectedRole={selectedRole || role}
-                onChange={setSelectedRole}
-              />
-            </div>
-            <Page.P>
-              The role defines the rights of a member of the workspace.{" "}
-              {ROLES_DATA[role]["description"]}
-            </Page.P>
-          </div>
-          <div className="flex flex-none flex-col gap-2">
-            <div className="flex-none">
-              <Button
-                variant="primaryWarning"
-                label="Revoke member access"
-                size="sm"
-                onClick={() => setRevokeMemberModalOpen(true)}
-              />
-            </div>
-            <Page.P>
-              Deleting a member will remove them from the workspace. They will
-              be able to rejoin if they have an invitation link.
-            </Page.P>
-          </div>
-        </div>
-      </Page>
-      <Dialog
-        isOpen={revokeMemberModalOpen}
-        title="Revoke member access"
-        onValidate={async () => {
-          await handleMembersRoleChange({
-            members: [member],
-            role: "none",
-            sendNotification,
-          });
-          await mutate(`/api/w/${owner.sId}/members`);
-          setRevokeMemberModalOpen(false);
-          onClose();
-        }}
-        validateLabel="Yes, revoke"
-        validateVariant="primaryWarning"
-        onCancel={() => {
-          setRevokeMemberModalOpen(false);
-        }}
-        isSaving={isSaving}
-      >
-        <div>
-          Revoke access for user{" "}
-          <span className="font-bold">{member.fullName}</span>?
-        </div>
-      </Dialog>
-    </ElementModal>
   );
 }
